requirements.txt
docx==0.2.4
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.22

